+++ bip_utils/conf/common/__init__.py
@@ -1,4 +1,4 @@
 from bip_utils.conf.common.bip_coin_conf_const import *
-from bip_utils.conf.common.bip_coin_conf_helper import CoinNames, KeyNetVersions
+from bip_utils.conf.common.bip_coin_conf_helper import KeyNetVersions
 from bip_utils.conf.common.bip_coin_conf_enum import AddrTypes, Bip32Types
 from bip_utils.conf.common.bip_coin_conf import *
