+++ bip_utils/utils/__init__.py
@@ -1,6 +1,7 @@
 from bip_utils.utils.algo import AlgoUtils
 from bip_utils.utils.base32 import Base32Encoder
 from bip_utils.utils.bit import BitUtils
+from bip_utils.utils.coin_names import CoinNames
 from bip_utils.utils.conversion import ConvUtils
 from bip_utils.utils.crypto import CryptoUtils
 from bip_utils.utils.data_bytes import DataBytes
